refactor(agents): keep bootstrap routing runtime-light
